Remove deprecated function call from LibJackWrapper. All assemblies are now version 0.2

+++ MonoMultiJack.ConnectionWrapper.Jack/LibJackWrapper.PInvoke.cs
@@ -48,7 +48,7 @@ namespace MonoMultiJack.ConnectionWrapper.Jack
 		
 		[DllImport(JACK_LIB_NAME)]
 		private static extern IntPtr jack_client_open(string client_name, 
-		                                             IntPtr jack_options_t, 
+		                                             byte jack_options_t, 
 		                                             IntPtr jack_status_t);
 
 		[DllImport(JACK_LIB_NAME)]
